Graph structure refactor
I changed representation of the graph to a more space efficient approach. The current structure stores a lot of redundant data as each step in the LinkedList stores a copy of that nodes data, not a pointer towards it.

I also changed the adjacency list into a dictionary as Cat wanted. I think this it is because searching for a node has a O(1) where the current approach needs to search the list, which has a O(n).

See the word document for details.