Fixed parse + Basic outline for populate flow
I fixed an issue with the parse CSV function as it didn't check if a node was currently in the graph before adding it.

I also made the populate flow_rate function actually set the flow rate. I still need to convert it to a proper queue and removing the testing info.